required auth of routes

+++ src/components/WorkoutDetails.js
@@ -1,15 +1,23 @@
-// import {WorkoutContext} from "../context/WorkoutContext";
-import {useWorkoutContext} from "../hooks/useWorkoutContext";
-
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+
+import { useWorkoutContext } from "../hooks/useWorkoutContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 export const WorkoutDetails = (props) => {
 	const workout = props.workout;
 	const { dispatch } = useWorkoutContext();
+	const { user } = useAuthContext();
 
 	const handleDeleteWorkout = async () => {
+		if (!user) {
+			return;
+		}
+
 		const response = await fetch('/api/workouts/' + workout._id, {
-			method: 'DELETE'
+			method: 'DELETE',
+			headers: {
+				'Authorization': `Bearer ${user.token}`
+			}
 		});
 		const json = await response.json();
 		if (response.ok) {
